+++ flytekit/core/data_persistence.py
@@ -298,7 +298,7 @@ class FileAccessProvider(object):
 
         self._default_remote = DataPersistencePlugins.find_plugin(raw_output_prefix)(default_prefix=raw_output_prefix)
         self._raw_output_prefix = raw_output_prefix
-        if not self.is_remote(raw_output_prefix):
+        if not self.is_remote(raw_output_prefix) and not os.path.exists(raw_output_prefix):
             os.mkdir(raw_output_prefix)
 
     @staticmethod
